상품 인가 기능 구현

+++ src/controllers/product.controller.js
@@ -3,6 +3,7 @@ const productService = require('../services/product.service');
 exports.getAllProducts = async (req, res, next) => {
     try {
         const result = await productService.getAllProducts(req.query);
+        console.log(`✅ 모든 상품 가져옴:`, result);
         res.status(200).json(result);
     } catch (err) {
         next(err);
@@ -12,6 +13,7 @@ exports.getAllProducts = async (req, res, next) => {
 exports.getProductById = async (req, res, next) => {
     try {
         const result = await productService.getProductById(req.params.id);
+        console.log(`✅ 개별 상품 가져옴:`, result);
         res.status(200).json(result);
     } catch (err) {
         next(err);
@@ -20,17 +22,29 @@ exports.getProductById = async (req, res, next) => {
 
 exports.createProduct = async (req, res, next) => {
     try {
-        const result = await productService.createProduct(req.body);
-        res.status(201).json(result);
-    } catch (err) {
-        next(err);
+        // createProduct 서비스에서 userId를 받아서 저장
+        const userId = req.user.id;
+        const result = await productService.createProduct({ ...req.body, userId });
+        console.log(`✅ 상품 등록 완료:`, result);
+        res.status(201).json({ message: 'Successfully created', result });
+    } catch (error) {
+        console.error(`error:`, error);
+        return next(err);
     }
 };
 
 exports.updateProduct = async (req, res, next) => {
     try {
+        const userId = req.user.id;
+        const product = await productService.getProductById(req.params.id);
+
+        if (Number(product.userId) !== Number(userId)) {
+            return res.status(403).json({ message: 'Unauthorized' });
+        }
+
         const result = await productService.updateProduct(req.params.id, req.body, req.file);
-        res.status(200).json(result);
+        console.log(`✅ 업데이트 완료:`, result);
+        res.status(200).json({ message: 'Successfully updated', result });
     } catch (err) {
         next(err);
     }
@@ -38,8 +52,16 @@ exports.updateProduct = async (req, res, next) => {
 
 exports.deleteProduct = async (req, res, next) => {
     try {
+        const userId = req.user.id;
+        const product = await productService.getProductById(req.params.id);
+
+        if (Number(product.userId) !== Number(userId)) {
+            return res.status(403).json({ message: 'Unauthorized' });
+        }
+
         const result = await productService.deleteProduct(req.params.id);
-        res.status(200).json(result);
+        console.log(`✅ 삭제 완료:`, result);
+        res.status(200).json({ message: 'Successfully deleted' });
     } catch (err) {
         next(err);
     }
